Move pymatics module handling out of definitions
