+++ mathics/builtin/assignments/assignment.py
@@ -5,6 +5,7 @@ Forms of Assignment
 
 
 from mathics.builtin.base import BinaryOperator, Builtin
+from mathics.builtin.pymathics import PyMathicsLoadException
 from mathics.core.assignment import (
     ASSIGNMENT_FUNCTION_MAP,
     AssignmentException,
@@ -18,7 +19,6 @@ from mathics.core.attributes import (
     A_PROTECTED,
     A_SEQUENCE_HOLD,
 )
-from mathics.core.definitions import PyMathicsLoadException
 from mathics.core.evaluators import eval_load_module
 from mathics.core.symbols import SymbolNull
 from mathics.core.systemsymbols import SymbolFailed
